refactor(spring): extract secret chain to beans + core-owned markers (Epic 6.3, step 2a)
Behavior-preserving split of the secret resolution chain out of the giant
tramai() bean into TramaiSecretResolutionAutoConfiguration, and introduction
of the two core-owned markers that make the module decomposition possible
without core knowing any provider or secret backend:

- SpringConfiguredModelProvider: provider modules register their constructed
  providers as these descriptors; core collects them as property-generated
  providers while plain ModelProvider beans remain user-supplied. Preserves
  bean-over-property precedence and duplicate detection exactly.
- SpringBuiltInSecretValueResolver: marker distinguishing built-in secret
  resolvers (Vault/AWS modules) from user resolvers, keeping the documented
  bootstrap-vs-full chain ordering.
- SpringSecretResolution: shared resolveSecret with byte-identical guard
  messages (both-key-and-ref, unresolvable-ref).
- tramai() now injects the full chain bean (@Qualifier tramaiSecretValueResolver).
- Chain beans back off when a user supplies their own Tramai bean (matches
  original inline-assembly semantics).

Characterization suite unchanged in assertions; every runner now registers
TramaiSecretResolutionAutoConfiguration (harness wiring only). 61/61 tests
green; verifyPr green.
