+++ tramai-spring-core/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,3 @@
+dev.tramai.spring.TramaiSecretResolutionAutoConfiguration
 dev.tramai.spring.TramaiAutoConfiguration
 dev.tramai.spring.SecurityClassificationAutoConfiguration
